asdfasdfasdf
-buttons change color depending on whether the upgrade can be afforded
-can upgrade cannons indefinitely
-hovering shop buttons
This may be as far as I will go with this project. If I ever want to expand on this project, I will do it with LibGDX or another legitimate game library. This project was mainly meant to see how far I could go with the limits of Java Swing and AWT without the aid of other external libraries.